requirements.txt
pillow
BIO
python-dotenv
gunicorn
dash
dash-bootstrap-components
